packages/image_to_2dgrid/pyproject.toml
[project]
name = "evo-data-converters-image-grid"
version = "0.1.0"
description = "Image Grid data converter for EVO (supports JPEG, PNG, TIFF, etc.)"
license-files = ["LICENSE.md"]
dynamic = ["readme"]
requires-python = ">=3.10"
authors = [
  { name = "Seequent", email = "support@seequent.com" }
]


dependencies = [
  "evo-data-converters-common==0.3.1",
  "evo-objects==0.3.3",
  "evo-schemas==2025.12.15",
  "evo-sdk-common==0.5.12",
  "numpy<=1.26.4",
  "pyarrow",
  "pydantic",
  "pillow",
]

[project.urls]
Source = "https://github.com/SeequentEvo/evo-data-converters"
Tracker = "https://github.com/SeequentEvo/evo-data-converters/issues"
Homepage = "https://www.seequent.com/"
Documentation = "https://developer.seequent.com/"

[dependency-groups]
dev = [
    "pytest",
    "requests-mock",
]

[tool.ruff]
src = ["src", "tests"]
line-length = 120

[build-system]
requires = ["hatchling", "hatch-fancy-pypi-readme"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/evo"]

[tool.hatch.metadata.hooks.fancy-pypi-readme]
content-type = "text/markdown"

[[tool.hatch.metadata.hooks.fancy-pypi-readme.fragments]]
path = "README.md"

[[tool.hatch.metadata.hooks.fancy-pypi-readme.substitutions]]
# Literal TOML strings (single quotes) need no escaping of backslashes.
# Converts relative links to absolute links in PyPI
pattern = '\[(.+?)\]\(((?!https?://)\S+?)\)'
replacement = '[\1](https://github.com/SeequentEvo/evo-data-converters/tree/main/packages/image_to_2dgrid/\g<2>)'
